Move popup code to its own module and remove various bits of dead code
