+++ src/Classes/LabelControl.lua
@@ -8,7 +8,7 @@ local LabelClass = newClass("LabelControl", "Control")
 
 ---@param anchor? ControlAnchor
 ---@param rect? ControlRect
----@param label string
+---@param label Prop<string>
 function LabelClass:LabelControl(anchor, rect, label)
 	self:Control(anchor, rect)
 	self.label = label
